update uploadphoto page layout,  change AllPhotosPosts name to AllPhotos, and delete PhotoPost file
